change scale factor to 16bit

+++ lib/src/kernels/transform/impl/mli_krn_softmax_ref.h
@@ -222,7 +222,7 @@ static mli_status mli_krn_softmax_sa8_run(const mli_tensor *in, const mli_softma
     struct s8asym_quant_params in_params;
     struct s8asym_quant_params out_params;
 
-    in_params.scale  = in->el_params.sa.scale.mem.i32;
+    in_params.scale  = in->el_params.sa.scale.mem.i16;
     in_params.shift = in->el_params.sa.scale_frac_bits.mem.i8;
     out_params.offset = kSoftmaxAsymZeroPoint;
     out_params.scale  = 1;
@@ -350,7 +350,7 @@ static mli_status mli_krn_softmax_sa8_run(const mli_tensor *in, const mli_softma
     }
 
     out->el_params.sa.zero_point.mem.i16 = out_params.offset;
-    out->el_params.sa.scale.mem.i32 = out_params.scale;
+    out->el_params.sa.scale.mem.i16 = out_params.scale;
     out->el_params.sa.scale_frac_bits.mem.i8 = (int8_t)out_params.shift;
 
     return MLI_STATUS_OK;
